requirements.txt
# requirement for running jupyter notebook in IntelliJ IDEA
jupyter

# used for encryption of data-files
python-dotenv
pyAesCrypt

# requirements for export
numpy
pandas
pymongo

# streamlit & dashboard
streamlit
plotly
matplotlib
seaborn

# ML Model
transformers
torch
iso-639

